chore: deprecate the vdp-sdk/metric in toolkit
