+++ packages/toolkit/src/lib/vdp-sdk/index.ts
@@ -4,7 +4,6 @@ import { env } from "../../server";
 import { Nullable } from "../type";
 
 export * from "./helper";
-export * from "./metric";
 export * from "./model";
 export * from "./operation";
 export * from "./organization";
