--- packages/toolkit/src/lib/vdp-sdk/metric/credit/types.ts
@@ -1,5 +1,0 @@
-export type CreditConsumptionChartRecord = {
-  creditOwner: string;
-  timeBuckets: string[];
-  amount: number[];
-};
