--- packages/toolkit/src/lib/vdp-sdk/metric/index.ts
@@ -1,2 +1,0 @@
-export * from "./pipeline";
-export * from "./credit";
