--- packages/toolkit/src/lib/vdp-sdk/metric/pipeline/index.ts
@@ -1,2 +1,0 @@
-export * from "./queries";
-export * from "./types";
